+++ malachite-q/src/test_util/gaussian_rational/arithmetic/mod.rs
@@ -10,5 +10,6 @@ pub mod add;
 pub mod div;
 pub mod mul;
 pub mod pow;
+pub mod root;
 pub mod sqrt;
 pub mod square;