+++ malachite-q/src/test_util/generators/random.rs
@@ -324,6 +324,28 @@ pub fn random_gaussian_rational_unsigned_pair_gen_var_1<T: PrimitiveUnsigned>(
     ))
 }
 
+pub fn random_gaussian_rational_unsigned_pair_gen_var_2<T: PrimitiveUnsigned>(
+    config: &GenConfig,
+) -> It<(GaussianRational, T)> {
+    Box::new(random_pairs(
+        EXAMPLE_SEED,
+        &|seed| {
+            random_gaussian_rationals(
+                seed,
+                config.get_or("mean_bits_n", 64),
+                config.get_or("mean_bits_d", 1),
+            )
+        },
+        &|seed| {
+            geometric_random_positive_unsigneds(
+                seed,
+                config.get_or("mean_small_n", 64),
+                config.get_or("mean_small_d", 1),
+            )
+        },
+    ))
+}
+
 // -- (GaussianRational, Rational) --
 
 pub fn random_gaussian_rational_rational_pair_gen(